Split up tile.cpp/h
