+++ src/client/content_cso.cpp
@@ -19,7 +19,6 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #include "content_cso.h"
 #include <IBillboardSceneNode.h>
-#include "client/tile.h"
 #include "clientenvironment.h"
 #include "client.h"
 #include "map.h"
