+++ src/client/mapblock_mesh.cpp
@@ -33,6 +33,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include <array>
 #include <algorithm>
 #include <cmath>
+#include "client/texturesource.h"
 
 /*
 	MeshMakeData
