+++ src/client/particles.h
@@ -21,7 +21,6 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #include <iostream>
 #include "irrlichttypes_extrabloated.h"
-#include "client/tile.h"
 #include "localplayer.h"
 #include "../particles.h"
