+++ src/client/texturepaths.h
@@ -1,0 +1,44 @@
+/*
+Minetest
+Copyright (C) 2010-2013 celeron55, Perttu Ahola <celeron55@gmail.com>
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation; either version 2.1 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+
+#pragma once
+
+#include <string>
+#include <vector>
+
+// Texture paths get cached and this clears the Cache.
+void clearTextureNameCache();
+
+// Find out the full path of an image by trying different filename extensions.
+// If failed, return "".
+std::string getImagePath(std::string path);
+
+/* Gets the path to a texture by first checking if the texture exists
+ * in texture_path and if not, using the data path.
+ *
+ * Checks all supported extensions by replacing the original extension.
+ *
+ * If not found, returns "".
+ *
+ * Utilizes a thread-safe cache.
+*/
+std::string getTexturePath(const std::string &filename, bool *is_base_pack = nullptr);
+
+// Returns all dictionaries found from the "texture_path" setting.
+std::vector<std::string> getTextureDirs();
