+++ src/gui/guiBackgroundImage.cpp
@@ -18,6 +18,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include "guiBackgroundImage.h"
 #include "client/guiscalingfilter.h"
 #include "log.h"
+#include "client/texturesource.h"
 
 GUIBackgroundImage::GUIBackgroundImage(gui::IGUIEnvironment *env,
 	gui::IGUIElement *parent, s32 id, const core::rect<s32> &rectangle,
