+++ src/gui/guiTable.cpp
@@ -28,7 +28,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "client/renderingengine.h"
 #include "debug.h"
 #include "log.h"
-#include "client/tile.h"
+#include "client/texturesource.h"
 #include "gettime.h"
 #include "util/string.h"
 #include "util/numeric.h"
