+++ src/nodedef.cpp
@@ -357,7 +357,7 @@ void ContentFeatures::reset()
 	has_on_destruct = false;
 	has_after_destruct = false;
 	floats = false;
-	
+
 	/*
 		Actual data
